Fix QGCView.showDialog() when called from FactTextField.
Actually, changed so showDialog figures out on its own if it should use available height or full parent height. The variable is now set by the main window, which controls which view is being shown.
I commented off some code in ParameterEditorDialog.qml as it is not properly handling the dialog and causing it to get "stuck".